+++ tests/testing.h
@@ -2,11 +2,12 @@
 #define _TESTING_H
 
 #include <stdio.h>
+#include <stdint.h>
 
 void equalchr(const char given, const char expected);
 void equalstr(const char *given, const char *expected);
-void equalnstr(const char *given, const char *expected, size_t len);
-void equalbin(const unsigned char *given, const unsigned char *, size_t len);
+void equalnstr(const char *given, const char *expected, uint32_t len);
+void equalbin(const unsigned char *given, const unsigned char *, uint32_t len);
 void equalint(int given, int expected);
